putting everything in a config file
